+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Check release notes for the 5.2.0-beta1 and beta2 below.
 ## Fixes
 
 - #333 - Fixes to method and service calls on interfaces/unions
+- #344 - Support paging field extensions with `field.ResolveBulk<TService, TKey, TResult>()`
 
 # 5.2.0-beta2